use args, update readme, delete run_bert_finetune.py
