+++ PaddleNLP/examples/distill/distill_lstm/bert_distill.py
@@ -1,3 +1,17 @@
+# Copyright (c) 2021 PaddlePaddle Authors. All Rights Reserve.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 import time
 
 import paddle
@@ -9,6 +23,7 @@ from paddlenlp.transformers.tokenizer_utils import whitespace_tokenize
 from paddlenlp.metrics import AccuracyAndF1
 from paddlenlp.datasets import GlueSST2, GlueQQP, ChnSentiCorp
 
+from args import parse_args
 from small import BiLSTM
 from data import create_distill_loader, load_embedding
 
@@ -20,24 +35,14 @@ TASK_CLASSES = {
 
 
 class Teacher(object):
-    def __init__(self,
-                 model_name='bert-base-uncased',
-                 param_path='model/SST-2/best_model_610/model_state.pdparams',
-                 max_seq_len=128):
-        self.max_seq_len = max_seq_len
+    def __init__(self, model_name, param_path):
         self.tokenizer = BertTokenizer.from_pretrained(model_name)
         self.model = BertForSequenceClassification.from_pretrained(model_name)
         self.model.set_state_dict(paddle.load(param_path))
         self.model.eval()
 
 
-def evaluate(task_name,
-             model,
-             ce_loss,
-             mse_loss,
-             metric,
-             data_loader,
-             alpha=0.0):
+def evaluate(task_name, model, metric, data_loader):
     model.eval()
     metric.reset()
     for i, batch in enumerate(data_loader):
@@ -61,83 +66,51 @@ def evaluate(task_name,
                 res[3],
                 res[4], ),
             end='')
-    elif isinstance(metric, Mcc):
-        print("mcc: %s, " % (res[0]), end='')
-    elif isinstance(metric, PearsonAndSpearman):
-        print(
-            "pearson: %s, spearman: %s, pearson and spearman: %s, " %
-            (res[0], res[1], res[2]),
-            end='')
     else:
         print("acc: %s, " % (res), end='')
     model.train()
 
 
-def do_train(
-        task_name='sst-2',
-        num_epoch=6,
-        batch_size=128,
-        opt='adadelta',
-        lr=1.0,
-        max_seq_length=128,
-        emb_dim=300,
-        hidden_size=300,
-        output_dim=2,
-        vocab_size=30522,
-        padding_idx=0,
-        dropout_prob=0.5,
-        save_steps=20,
-        alpha=0.0,
-        model_name='bert-base-uncased',
-        teacher_path='model/SST-2/best_model_610/model_state.pdparams',
-        vocab_path='/root/.paddlenlp/models/bert-base-uncased/bert-base-uncased-vocab.txt',
-        use_pretrained_w2v=True,
-        data_augmentation=False,
-        n_iter=20):
+def do_train(agrs):
     train_data_loader, dev_data_loader = create_distill_loader(
-        task_name,
-        model_name=model_name,
-        vocab_path=vocab_path,
-        batch_size=batch_size,
-        max_seq_length=max_seq_length,
-        data_augmentation=data_augmentation,
-        n_iter=n_iter)
+        args.task_name,
+        model_name=args.model_name,
+        vocab_path=args.vocab_path,
+        batch_size=args.batch_size,
+        max_seq_length=args.max_seq_length,
+        n_iter=args.n_iter)
 
     emb_tensor = load_embedding(
-        vocab_path=vocab_path) if use_pretrained_w2v else None
-    model = BiLSTM(
-        emb_dim,
-        hidden_size,
-        vocab_size,
-        output_dim,
-        padding_idx,
-        dropout_prob=dropout_prob,
-        embed_weight=emb_tensor)
+        args.vocab_path) if args.use_pretrained_emb else None
 
-    if opt == 'adadelta':
+    model = BiLSTM(args.emb_dim, args.hidden_size, args.vocab_size,
+                   args.output_dim, args.padding_idx, args.num_layers,
+                   args.dropout_prob, args.init_scale, emb_tensor)
+
+    if args.optimizer == 'adadelta':
         optimizer = paddle.optimizer.Adadelta(
-            learning_rate=lr, rho=0.95, parameters=model.parameters())
+            learning_rate=args.lr, rho=0.95, parameters=model.parameters())
     else:
         optimizer = paddle.optimizer.Adam(
-            learning_rate=lr, parameters=model.parameters())
+            learning_rate=args.lr, parameters=model.parameters())
 
     ce_loss = nn.CrossEntropyLoss()
     mse_loss = nn.MSELoss()
     klloss = nn.KLDivLoss()
 
-    metric_class = TASK_CLASSES[task_name][1]
+    metric_class = TASK_CLASSES[args.task_name][1]
     metric = metric_class()
 
-    teacher = Teacher(model_name=model_name, param_path=teacher_path)
+    teacher = Teacher(model_name=args.model_name, param_path=args.teacher_path)
 
-    print("Start to train distilling model.")
+    print("Start to distill small model.")
 
     global_step = 0
     tic_train = time.time()
-    for epoch in range(num_epoch):
+    for epoch in range(args.max_epoch):
         model.train()
         for i, batch in enumerate(train_data_loader):
-            if task_name == 'qqp':
+            if args.task_name == 'qqp':
                 bert_input_ids, bert_segment_ids, small_input_ids_1, seq_len_1, small_input_ids_2, seq_len_2, labels = batch
             else:
                 bert_input_ids, bert_segment_ids, small_input_ids, seq_len, labels = batch
@@ -147,27 +120,26 @@ def do_train(
                 teacher_logits = teacher.model(bert_input_ids, bert_segment_ids)
 
             # Calculate student model's forward.
-            if task_name == 'qqp':
+            if args.task_name == 'qqp':
                 logits = model(small_input_ids_1, seq_len_1, small_input_ids_2,
                                seq_len_2)
             else:
                 logits = model(small_input_ids, seq_len)
 
-            loss = alpha * ce_loss(logits, labels) + (1 - alpha) * mse_loss(
-                logits, teacher_logits)
+            loss = args.alpha * ce_loss(logits, labels) + (
+                1 - args.alpha) * mse_loss(logits, teacher_logits)
 
             loss.backward()
             optimizer.step()
             optimizer.clear_grad()
 
-            if i % save_steps == 0:
+            if i % args.log_freq == 0:
                 print(
                     "global step %d, epoch: %d, batch: %d, loss: %f, speed: %.4f step/s"
                     % (global_step, epoch, i, loss,
-                       save_steps / (time.time() - tic_train)))
+                       args.log_freq / (time.time() - tic_train)))
                 tic_eval = time.time()
-                acc = evaluate(task_name, model, ce_loss, mse_loss, metric,
-                               dev_data_loader, alpha)
+                acc = evaluate(args.task_name, model, metric, dev_data_loader)
                 print("eval done total : %s s" % (time.time() - tic_eval))
                 tic_train = time.time()
             global_step += 1
@@ -175,63 +147,7 @@ def do_train(
 
 if __name__ == '__main__':
     paddle.seed(2021)
-    # paddle.seed(202)
-    # task_name = 'senta'
-    task_name = 'sst-2'
-    # task_name = 'qqp'
+    args = parse_args()
+    print(args)
 
-    vocab_size_dict = {
-        "senta": 29496,
-        # "bert-base-chinese": 21128,
-        # "bert-base-uncased": 30522,
-        "sst-2": 30522,
-        "qqp": 30522,
-    }
-    base_teacher_path_dict = {
-        "sst-2": 'model/SST-2/best_model_610/model_state.pdparams',
-        "qqp": "model/QQP/best_model_17000/model_state.pdparams",
-        # "senta": 'model/senta/best_bert_base_model_930/model_state.pdparams',
-        "senta": 'model/senta/best_bert_wwm_ext_model_880/model_state.pdparams',
-    }
-    large_teacher_path_dict = {
-        "sst-2": 'model_large/SST-2/best_model_9450/model_state.pdparams',
-    }
-    if task_name == 'senta':
-        do_train(
-            task_name=task_name,
-            alpha=0.0,
-            dropout_prob=0.1,
-            lr=1.0,
-            data_augmentation=True,
-            batch_size=128,
-            num_epoch=20,
-            vocab_size=vocab_size_dict[task_name],
-            model_name='bert-base-chinese',
-            teacher_path=base_teacher_path_dict[task_name],
-            use_pretrained_w2v=False,
-            vocab_path='senta_word_dict_subset.txt')
-
-    elif task_name == 'sst-2':
-        do_train(
-            task_name=task_name,
-            alpha=0.0,
-            dropout_prob=0.3,
-            data_augmentation=True,
-            batch_size=128,
-            vocab_size=vocab_size_dict[task_name],
-            model_name='bert-base-uncased',
-            teacher_path=base_teacher_path_dict[task_name],
-            use_pretrained_w2v=True)
-    else:  # qqp
-        do_train(
-            task_name=task_name,
-            alpha=0.0,
-            dropout_prob=0.2,
-            num_epoch=20,
-            data_augmentation=True,
-            batch_size=256,
-            vocab_size=vocab_size_dict[task_name],
-            model_name='bert-base-uncased',
-            teacher_path=base_teacher_path_dict[task_name],
-            use_pretrained_w2v=True,
-            n_iter=10)
+    do_train(args)
